Avances historias de usuario

+++ easyevent/src/main/java/com/ucatolica/easyevent/easyevent/repository/ProveedorRepository.java
@@ -9,6 +9,6 @@ import java.io.Serializable;
 import java.util.List;
 import java.util.Optional;
 @Repository
-public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {
+public interface ProveedorRepository extends JpaRepository<Proveedor,Integer> {
 
 }
